moved slider handling to slider manager
