+++ Scripts/View/ComponentViews/ComponentView.cs
@@ -32,7 +32,6 @@ namespace ConnectAPIC.LayoutWindow.View
     {
         public int WidthInTiles { get; set; }
         public int HeightInTiles { get; set; }
-        private List<Godot.Slider> Sliders { get; set; } = new();
         private Node2D RotationArea { get; set; } // the part of the component that rotates
         public Sprite2D OverlayBluePrint { get; set; }
         public Sprite2D OverlayRed { get; private set; } // each laser(color) is independent of the others, so they need their own overlay and shader
@@ -43,9 +42,6 @@ namespace ConnectAPIC.LayoutWindow.View
         private List<Sprite2D> OverlaySprites { get; set; } = new();
         public ShaderMaterial LightOverlayShader { get; set; }
         public ComponentViewModel ViewModel { get; private set; }
-        public const string SliderNumberMetaID = "SliderNumber";
-        public const string SliderLabelMetaID = "SliderLabel";
-        public const string SliderIsCallbackRegistered = "SliderIsCallbackRegistered";
         public new float RotationDegrees{
             get => RotationArea?.RotationDegrees ?? 0;
             set{
@@ -54,21 +50,13 @@ namespace ConnectAPIC.LayoutWindow.View
             }
         }
         private new float Rotation { get => RotationArea.Rotation; set => RotationArea.Rotation = value; }
+        private SliderManager sliderManager { get; set; }
 
         public ComponentView()
         {
             ViewModel = new ComponentViewModel();
             ViewModel.PropertyChanged += ViewModel_PropertyChanged;
-            ViewModel.SliderData.CollectionChanged += (object sender, NotifyCollectionChangedEventArgs e) =>
-            {
-                if(e.Action == NotifyCollectionChangedAction.Add)
-                {
-                    foreach ( var slider in e.NewItems.Cast<SliderViewData>().ToList())
-                    {
-                        FindAndInitializeSlider(slider);
-                    }
-                }
-            };
+            sliderManager = new SliderManager(ViewModel, this);
         }
 
         private void ViewModel_PropertyChanged(object sender, System.ComponentModel.PropertyChangedEventArgs e)
@@ -154,40 +142,7 @@ namespace ConnectAPIC.LayoutWindow.View
                 new (LaserType.Blue,tileOffset, inflowSide, OverlayBlue, overlayAnimTexture, new Godot.Vector2I(WidthInTiles, HeightInTiles)),
             };
         }
-        private void SetSliderValue(int sliderNumber, double value)
-        {
-            var slider = Sliders.Single(s => (int)s.GetMeta(SliderNumberMetaID) == sliderNumber);
-            var label = (RichTextLabel)slider.GetMeta(SliderLabelMetaID);
-            slider.Value = value;
-            SetSliderLabelText(label, value);
-        }
-
-        // initialize one of the existing sliders
-        private void FindAndInitializeSlider(SliderViewData sliderData)
-        {
-            var label = FindChild(sliderData.GodotSliderLabelName, true, false) as RichTextLabel;
-            var godotSlider = FindChild(sliderData.GodotSliderName, true, false) as Godot.Slider;
-            godotSlider.MinValue = sliderData.MinVal;
-            godotSlider.MaxValue = sliderData.MaxVal;
-            godotSlider.SetMeta(SliderNumberMetaID, sliderData.Number);
-            godotSlider.SetMeta(SliderLabelMetaID, label);
-            if((bool)godotSlider.GetMeta(SliderIsCallbackRegistered) != true)
-            {
-                godotSlider.SetMeta(SliderIsCallbackRegistered, true);
-                godotSlider.ValueChanged += (newVal) =>
-                {
-                    SetSliderLabelText(label, newVal);
-                    ViewModel.SetSliderValue(sliderData.Number, newVal);
-                };
-            }
-
-            godotSlider.Value = sliderData.Value;
-            SetSliderLabelText(label, sliderData.Value);
-            godotSlider.Step = (sliderData.MaxVal - sliderData.MinVal) / sliderData.Steps; // step is the distance between two steps in value
-            this.Sliders.Add(godotSlider);
-        }
-
-        private void SetSliderLabelText(RichTextLabel label, double newVal) => label.Text = $"[center]{newVal:F2}";
+        
         private void FindAndCreateLightOverlays()
         {
             FindOverlayBlueprint();
@@ -330,30 +285,12 @@ namespace ConnectAPIC.LayoutWindow.View
             copy.ViewModel.RotationCC = ViewModel.RotationCC; // give the new copy the proper RotationCC so that it has the correct rotation
 
             // deep copy that list of sliders
-            List<SliderViewData> newSliderData = DuplicateSliders();
+            List<SliderViewData> newSliderData = sliderManager.DuplicateSliders();
             copy.ViewModel.InitializeComponent(ViewModel.TypeNumber, newSliderData, ViewModel.Logger);
             return copy;
         }
 
-        private List<SliderViewData> DuplicateSliders()
-        {
-            var newSliderData = new List<SliderViewData>();
-            foreach (var slider in ViewModel.SliderData)
-            {
-                newSliderData.Add(new SliderViewData()
-                {
-                    GodotSliderLabelName = slider.GodotSliderLabelName,
-                    GodotSliderName = slider.GodotSliderName,
-                    Value = slider.Value,
-                    MaxVal = slider.MaxVal,
-                    MinVal = slider.MinVal,
-                    Number = slider.Number,
-                    Steps = slider.Steps
-                });
-            }
-
-            return newSliderData;
-        }
+        
 
         public override void _ExitTree()
         {
